Move modal manager into a modal manager directory
