+++ packages/donation-form/demo/donate.html
@@ -66,7 +66,7 @@
     import { BraintreeManager } from '../lib/braintree-manager/braintree-manager.js';
     import { RecaptchaManager } from '../lib/recaptcha-manager/recaptcha-manager.js';
 
-    import '../lib/modals/modal-manager.js';
+    import '../lib/modal-manager/modal-manager.js';
     import '../lib/donation-form.js';
 
     const hostedFieldStyle = {
