+++ packages/donation-form/src/donation-form.ts
@@ -14,9 +14,9 @@ import './form-elements/form-section';
 import './form-elements/header/donation-form-header';
 import './form-elements/contact-form';
 import './form-elements/payment-selector';
-import './modals/modal-template';
+import './modal-manager/modal-template';
 import { BraintreeManagerInterface } from './braintree-manager/braintree-manager';
-import { ModalManagerInterface } from './modals/modal-manager';
+import { ModalManagerInterface } from './modal-manager/modal-manager';
 import { DonationRequest } from './models/request_models/donation-request';
 import { ContactForm } from './form-elements/contact-form';
 import { RecaptchaManagerInterface } from './recaptcha-manager/recaptcha-manager';
