+++ packages/donation-form/src/form-elements/payment-selector.ts
@@ -13,7 +13,7 @@ import {
 import { BraintreeManagerInterface } from '../braintree-manager/braintree-manager';
 import { DonationPaymentInfo } from '../models/donation-info/donation-payment-info';
 import { DonationFrequency } from '../models/donation-info/donation-frequency';
-import { ModalManagerInterface } from '../modals/modal-manager';
+import { ModalManagerInterface } from '../modal-manager/modal-manager';
 import { PayPalFlowHandler } from '../payment-flow-handlers/paypal-flow-handler';
 
 @customElement('payment-selector')
